Authentication 코드 리뷰 &  이론 정리 |  login_required 구현 | 회원 탈퇴 기능 구현!

+++ Django/05_django_form/articles/views.py
@@ -1,4 +1,5 @@
 from django.views.decorators.http import require_POST
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
@@ -6,13 +7,14 @@ from IPython import embed
 
 # Create your views here.
 def index(request):
+    # embed()
     articles = Article.objects.all()
     context = {
         'articles' : articles,
     }
     return render(request, 'articles/index.html', context)
 
-
+@login_required
 def create(request):
     if request.method=="POST":
         # embed()
@@ -60,10 +62,12 @@ def detail(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)
+        article.delete()
     return redirect('articles:index')
 
+@login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
 
@@ -92,23 +96,24 @@ def update(request, article_pk):
 @require_POST
 def comments_create(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    
-    comment_form = CommentForm(request.POST)
+    if request.user.is_authenticated:
+        
+        comment_form = CommentForm(request.POST)
 
-    if comment_form.is_valid():
-        # save 메서드 -> 선택 인자 : (기본 값) commit=True : DB에 바로 저장
-        # commit=False : DB에 바로 저장되는 것을 막아준다.
-        # 객체 저장 
-        comment = comment_form.save(commit=False)
-        comment.article = article
-        comment.save()
+        if comment_form.is_valid():
+            # save 메서드 -> 선택 인자 : (기본 값) commit=True : DB에 바로 저장
+            # commit=False : DB에 바로 저장되는 것을 막아준다.
+            # 객체 저장 
+            comment = comment_form.save(commit=False)
+            comment.article = article
+            comment.save()
 
-    return redirect('articles:detail', article.pk)
+    return redirect('articles:detail', article_pk)
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
-    article = get_object_or_404(Article, pk=article_pk)
-    
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()    
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=article_pk)        
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        comment.delete()    
     return redirect('articles:detail', article.pk)
